+++ views/homepage.handlebars
@@ -1,2 +1,1 @@
 <p>This is the homepage</p>
-
